% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/an_cont.R
\name{an_cont}
\alias{an_cont}
\title{Analyse a continuous outcome in a cross-over dataset}
\usage{
an_cont(data, outcome, method = "lmm", alternative = "two_sided",
  complete_case = T, conf_level = 0.95, carryover = F, summary = T,
  ...)
}
\arguments{
\item{data}{A \code{\link[tibble]{tibble}} of class \code{xover_data}.}

\item{outcome}{The chosen continuous outcome variable from \code{data} for
which the analysis will be performed. Should be stated absolutely, i.e., not
as a \code{\link[base]{character}} string.}

\item{method}{A \code{\link[base]{character}} string indicating which method
to use for analysis. Must be one of \code{"anova"}, \code{"lmm"}, and
\code{"t_test"}.}

\item{alternative}{\code{\link[base]{character}} string indicating whether to
perform a one- or two-sided test. Must be one of \code{"one_sided"} and
\code{"two_sided"}.}

\item{complete_case}{A \code{\link[base]{logical}} variable indicating
whether to reduce \code{data} to its complete case form.}

\item{conf_level}{The confidence level to use in confidence interval
construction. Should be a single \code{\link[base]{numeric}} strictly between
zero and one.}

\item{carryover}{Only used in the case that \code{method} is
\code{"lmm"}, when it indicates whether carryover effects should be included
in the model. Should be a \code{\link[base]{logical}} variable.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
A \code{\link[base]{list}} or class \code{xover_an} containing in
particular a \code{\link[base]{list}} \code{test_results} describing the
results of the chosen test.
}
\description{
Facilitates the analysis of continuous outcome data in a cross-over dataset.
}
\details{
\code{an_cont()} supports the analysis of continuous outcome data from a
cross-over trial. Precisely, a dataset of class \code{xover_data} is provided
(see \code{data}), along with a continuous outcome variable within this
dataset (see \code{outcome}). The value of \code{method}, and several other
input arguments, then determines the type of analysis performed on the data.
At present, the following values of \code{method} are supported:

\itemize{
\item \code{"anova"}: To analyse using an analysis of variance approach;
\item \code{"lmm"}: To analyse using a linear mixed model;
\item \code{"t_test"}: To analyse using t-tests.
}

See the package vignette for further dettails.

Note that datasets can be converted to class \code{xover_data} using
\code{\link[xover]{as_xover_data}}.
}
\references{
Jones B, Kenward MG (2014) \emph{Design and Analysis of
Cross-Over Trials}. Chapman and Hall: London, 3rd Edition.
}
\seealso{
\code{\link[xover]{as_xover_data}} for converting datasets to class
\code{xover_data}.
}
