% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_xover_data.R
\name{as_xover_data}
\alias{as_xover_data}
\title{Convert dataset to class \code{xover_data}}
\usage{
as_xover_data(data, summary = T)
}
\arguments{
\item{data}{A \code{\link[base]{matrix}} or \code{\link[base]{data.frame}}
containing data corresponding to a cross-over trial.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
A \code{\link[tibble]{tibble}} of class \code{xover_data}.
}
\description{
Converts an input dataset to be of class \code{xover_data}.
}
\details{
\code{as_xover_data()} supports the conversion of an input dataset to be of
class \code{xover_data}. Precisely, a \code{\link[base]{matrix}} or
\code{\link[base]{data.frame}} is provided (see \code{data}) where each row
corresponds to a particular outcome variable, and the function
converts it to a \code{\link[tibble]{tibble}} with class \code{xover_data},
augmenting the column names as required. In adddition, checks are performed
that the dataset corresponds to a cross-over trial.

The input dataset must contain information on the period, subject, and
treatment/sequence for each outcome variable.
}
