% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_case_data.R
\name{complete_case_data}
\alias{complete_case_data}
\title{Complete case data}
\usage{
complete_case_data(data, outcome, summary = T)
}
\arguments{
\item{data}{A \code{\link[tibble]{tibble}} of class \code{xover_data}.}

\item{outcome}{The chosen outcome variable from \code{data} upon which the
complete case reduction will be based. Should be stated absolutely, i.e., not
as a \code{\link[base]{character}} string.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
A \code{\link[tibble]{tibble}} of class \code{xover_data}.
Specifically, the input dataset \code{data} reduced to its complete case
form.
}
\description{
Reduces a cross-over dataset to complete case form.
}
\details{
\code{complete_case_data()} supports the reduction of a cross-over dataset
to those outcomes corresponding to subjects with no missing data. Precisely,
a dataset of class \code{xover_data} is provided (see \code{data}), along
with an outcome variable within this dataset (see \code{outcome}), and the
rows corresponding to subjects for whom at least one outcome variable is
missing for the chosen outcome variable, based on an assessment of
\code{data$sequence}, are removed.

Note that datasets can be converted to class \code{xover_data} using
\code{\link[xover]{as_xover_data}}.
}
\references{
Jones B, Kenward MG (2014) \emph{Design and Analysis of
Cross-Over Trials}. Chapman and Hall: London, 3rd Edition.
}
\seealso{
\code{\link[xover]{as_xover_data}} for converting datasets to class
\code{xover_data}.
}
