% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table_data.R
\name{contingency_table_data}
\alias{contingency_table_data}
\title{Cross-over data contingency table}
\usage{
contingency_table_data(data, outcome, summary = T)
}
\arguments{
\item{data}{A \code{\link[tibble]{tibble}} of class \code{xover_data}.}

\item{outcome}{The chosen binary outcome variable from \code{data} upon which
the contingency_table will be based. Should be stated absolutely, i.e., not
as a \code{\link[base]{character}} string.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
A \code{\link[tibble]{tibble}} containing the determined contingency
table.
}
\description{
Builds a contingency table for a cross-over dataset.
}
\details{
\code{contingency_table_data()} supports the calculation of a contingency
table for a cross-over dataset. Precisely, a dataset of class
\code{xover_data} is provided (see \code{data}), along with a binary outcome
variable within this dataset (see \code{outcome}), and the data is reduced to
its complete case form for this outcome. A contingency table is then
constructed for this complete case data.

Note that datasets can be converted to class \code{xover_data} using
\code{\link[xover]{as_xover_data}}.
}
\references{
Jones B, Kenward MG (2014) \emph{Design and Analysis of
Cross-Over Trials}. Chapman and Hall: London, 3rd Edition.
}
\seealso{
\code{\link[xover]{as_xover_data}} for converting datasets to class
\code{xover_data}.
}
