% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cow_1.R
\docType{data}
\name{data_cow_1}
\alias{data_cow_1}
\title{A three-treatment crossover trial on milk-yield}
\format{A \code{\link[tibble]{tibble}} with 18 rows (observations) and 7
columns (variables):
\describe{
  \item{`milk yield`}{A continuous outcome measurement. Stored as an
  \code{\link[base]{integer}}.}
  \item{period}{The period of the trial corresponding to each observation.
  Stored as a \code{\link[base]{factor}}.}
  \item{subject}{The subject corresponding to each observation. Stored as a
  \code{\link[base]{factor}}.}
  \item{treatment}{The administered treatment corresponding to each
  observation. Stored as a \code{\link[base]{factor}}.}
  \item{sequence}{The treatment sequence for the subject corresponding to
  each observation. Stored as a \code{\link[base]{factor}}.}
  \item{`sequence index`}{The treatment sequence index for the subject
  corresponding to each observation. Stored as a \code{\link[base]{factor}}.}
  \item{square}{A categorical covariate. Stored as a
  \code{\link[base]{factor}}.}
}}
\source{
Ratkowsky DA, Evans MA, Alldredge JR (1992) \emph{Cross-Over
Experiments: Design, Analysis and Application}. Marcel Dekker: New York.
}
\usage{
data_cow_1
}
\description{
Quoted from Ratkowsky \emph{et al.} (1992) (see page 2): "Consider the pair
of \ifelse{html}{\out{3 &times; 3}}{\eqn{3 \times 3}} Latin squares given in
Table 1.1, taken from Cochran and Cox (1957), p. 135. The response variable
is milk yield. This pair of squares is an example of a "diagram-balanced"
pair, in which each treatment is preceded and followed equally often by each
other treatment."
}
\details{
Data pertains to a trial with

\itemize{
\item 3 treatments;
\item 3 periods;
\item 6 sequences;
\item 6 subjects;
\item 1 continuous outcome measurement;
\item 1 categorical covariate.
}
}
\keyword{datasets}
