% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_carryover_balanced.R
\name{is_carryover_balanced}
\alias{is_carryover_balanced}
\title{Check for carryover balance}
\usage{
is_carryover_balanced(sequences, pre_period = F, summary = T)
}
\arguments{
\item{sequences}{A cross-over design; an object of class \code{xover_seq}.}

\item{pre_period}{A \code{\link[base]{logical}} variable indicating whether a
pre-period should be assumed. See \strong{Details} for further information.
Defaults to \code{F}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
A \code{\link[base]{logical}} variable indicating whether the design
is balanced for first order carryover effects.
}
\description{
Checks whether a cross-over design is balanced for first order carryover
effects.
}
\details{
\code{is_carryover_balanced()} supports the assessment of whether a
cross-over design of class \code{xover_seq} (see \code{sequences}) is
balanced for first order carryover effects (residual effects). The user is
also able to specify whether there is a pre-period (see \code{pre_period}).
That is, if there is a pre-period each subject experiences in the first
period the residual effect of the treatment of the last period (i.e., the
last period preceeds the first period).

Ultimately, a cross-over design is said to be balanced for first order
carryover effects if each treatment is preceeded by all other treatments
equally often, and if no treatment is preceeded by itself.

Note that a cross-over design can be converted to be of class
\code{xover_seq} using \code{\link[xover]{as_xover_seq}}.
}
\examples{
# Check a three-treatment Latin square for carryover balance
latin    <- seq_latin(D = 3)
is_carryover_balanced(latin)
# Check a three-treatment Williams design for carryover balance
williams <- seq_williams(D = 3)
is_carryover_balanced(williams)
}
\seealso{
\code{\link[xover]{as_xover_seq}} for converting cross-over designs
to class \code{xover_seq}.
}
\author{
Based on code from the \code{\link[crossdes]{crossdes}} package by
Oliver Sailer.
}
