% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_link.R
\name{mat_link}
\alias{mat_link}
\title{Link matrix specification}
\usage{
mat_link(D = 4, model = 1, prop = 0.5, placebos = 1, summary = T)
}
\arguments{
\item{D}{The number of treatments. Must be a single
\code{\link[base]{numeric}} integer greater than or equal to two. Defaults to
\code{2}.}

\item{model}{A single \code{\link[base]{numeric}} integer between one and
nine inclusive, specifying which model to compute the row-column design for.
See \strong{Details} for further information. Defaults to \code{1}.}

\item{prop}{Only used when \code{model = 3}, when it should be a single
\code{\link[base]{numeric}} describing the proportionality parameter.}

\item{placebos}{Only used when \code{model = 4}, when it should be a single
\code{\link[base]{numeric}} integer between zero and \code{D - 1} inclusive,
describing the number of placebos in the trial.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
A \code{\link[base]{matrix}} detailing the determined link matrix.
}
\description{
Specifies the link matrix corresponding to a particular analysis model.
}
\details{
\code{mat_link()} supports the specification of link matrices. Precisely,
a number of treatments (see \code{D}) and a chosen model (see \code{model})
are specified. The link matrix for use in computing design matrices is then
returned. The value of \code{model} must be an integer between one and nine
inclusive, corresponding to the following models:

\itemize{
\item 1. Standard additive model;
\item 2. Second-order carry-over effects;
\item 3. Full-set of interactions;
\item 4. Self-adjacency model;
\item 5. Placebo model;
\item 6. No carry-over into self model;
\item 7. Treatment decay model;
\item 8. Proportionality model;
\item 9. No carry-over effects.
}

See the package vignette for further details on these models.
}
\author{
Based on code from the \code{\link[Crossover]{Crossover}} package by
Kornelius Rohmeyer.
}
