% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_rcd.R
\name{mat_rcd}
\alias{mat_rcd}
\title{Create the design matrix for a given row column design}
\usage{
mat_rcd(rcd, D = 2, model = 1)
}
\arguments{
\item{model}{String or number describing the model. See \code{\link{getModelNr}}.}

\item{X}{row-column design}

\item{v}{number of treatments}
}
\value{
The design matrix for a row-column design.
}
\description{
Create the design matrix for a given row column design
}
\seealso{
\code{\link{rcd}} gives the row-column design to a given crossover design.
}
