% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_seq.R
\name{order_seq}
\alias{order_seq}
\title{Sort cross-over design sequences}
\usage{
order_seq(sequences, summary = T)
}
\arguments{
\item{sequences}{An object of \code{\link[base]{class}} \code{xover_seq}; a
set of cross-over design sequences.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
An object of \code{\link[base]{class}} \code{xover_seq}, with
canonically ordered sequences.
}
\description{
Sorts a set of cross-over design sequences in to a canonical order.
}
\details{
Accepts an object of \code{\link[base]{class}} \code{xover_seq} (see
\code{sequences}), for which it will then sort the sequences in to a
canonical order (i.e., alphabetical or lexographical as appropriate).

When comparing cross-over designs with a larger number of sequences or
periods this ordering eases the ability to check whether two sequences are
equal.
}
\seealso{
The \code{seq_#()} functions within \code{\link[xover]{xover}}.
}
\author{
Based on code from the \code{\link[Crossover]{Crossover}} package by
Kornelius Rohmeyer.
}
