% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcd_seq.R
\name{rcd_seq}
\alias{rcd_seq}
\title{Row-column design specification}
\usage{
rcd_seq(sequences, model = 1, summary = T)
}
\arguments{
\item{sequences}{An object of class \code{xover_seq}, describing the
cross-over design for which the row-column design will be computed.}

\item{model}{A single \code{\link[base]{numeric}} integer between one and
nine inclusive, specifying which model to compute the row-column design for.
See \strong{Details} for further information. Defaults to \code{1}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
A \code{\link[base]{matrix}} of class \code{xover_rcd} detailing the
determined row-column design.
}
\description{
Specifies the row-column design corresponding to an input cross-over design.
}
\details{
\code{rcd_seq()} supports the specification of row-column designs. Precisely,
a crossover design is provided (see \code{sequences}) and the corresponding
row-column design for the chosen model (see \code{model}) is returned.
The value of \code{model} must be an integer between one and nine inclusive,
corresponding to the following models:

\itemize{
\item 1. Standard additive model;
\item 2. Second-order carry-over effects;
\item 3. Full-set of interactions;
\item 4. Self-adjacency model;
\item 5. Placebo model;
\item 6. No carry-over into self model;
\item 7. Treatment decay model;
\item 8. Proportionality model;
\item 9. No carry-over effects.
}

See the package vignette for further details on these models.
}
\examples{
# Row-column design for a three-treatment three-period Latin square
latin <- seq_latin(D = 3)
rcd   <- rcd_seq(latin)
}
\author{
Based on code from the \code{\link[Crossover]{Crossover}} package by
Kornelius Rohmeyer.
}
