% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_all.R
\name{seq_all}
\alias{seq_all}
\title{Supported cross-over designs}
\usage{
seq_all(D = 2, J = 2, K = 2, summary = T)
}
\arguments{
\item{D}{The number of treatments. Must be a single
\code{\link[base]{numeric}} integer greater than or equal to two. Defaults to
\code{2}.}

\item{J}{The number of periods. Must be a \code{\link[base]{numeric}} vector
of integers, whose values are greater than or equal to two. Defaults to
\code{2}.}

\item{K}{The number of sequences. Must be a \code{\link[base]{numeric}}
vector of integers, whose values are greater than or equal to two. Defaults
to \code{2}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
If at least one supported design is identified for the chosen input
parameters, a \code{\link[tibble]{tibble}} describing the supported
sequences. Otherwise, it will be \code{\link[base]{NULL}}. In either case,
the returned object will have class \code{xover_seq}.
}
\description{
Determine the cross-over designs supported by the \code{seq_#()} functions in
\code{\link[xover]{xover}}.
}
\details{
\code{seq_all()} supports the determination of the cross-over designs
supported by the \code{seq_#()} functions in \code{\link[xover]{xover}} for a
specified number of treatments (see \code{D}), periods (see \code{J}), and
sequences (see \code{K}).
}
\examples{
# Available sequences for three-treatment three-period trials with between
# two and 20 sequences
three_by_three        <- seq_all(D = 3, J = 3, K = 2:20)
# Available sequences for three-treatment two-period trials with between
# two and 20 sequences
three_by_two          <- seq_all(D = 3, J = 2, K = 2:20)
# Available sequences for three-treatment two- or three-period trials with
between two and 20 sequences
three_by_two_or_three <- seq_all(D = 3, J = c(2, 3), K = 2:20)
}
