% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_anderson.R
\name{seq_anderson}
\alias{seq_anderson}
\title{Anderson (2002) cross-over design specification}
\usage{
seq_anderson(labels = 0:6, as_matrix = T, summary = T)
}
\arguments{
\item{labels}{A \code{\link[base]{vector}} of labels for the treatments.
Should be of \code{\link[base]{length}} seven, containing unique elements.
Defaults to \code{0:6}.}

\item{as_matrix}{A \code{\link[base]{logical}} variable indicating whether
the design should be returned as a \code{\link[base]{matrix}}, or a
\code{\link[tibble]{tibble}}. Defaults to \code{T}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
Either a \code{\link[base]{matrix}} if \code{as_matrix = T} (with
rows corresponding to sequences and columns to periods), or a
\code{\link[tibble]{tibble}} if \code{as_matrix = F} (with rows corresponding
to a particular period on a particular sequence). In either case, the
returned object will have class \code{xover_seq}.
}
\description{
Specifies a cross-over design from Anderson (2002).
}
\details{
\code{seq_anderson()} supports the specification of a cross-over design from
Anderson (2002). Sequences are supported for any chosen treatment labels (see
\code{labels}). In addition, the design can be returned in
\code{\link[base]{matrix}} or \code{\link[tibble]{tibble}} form (see
\code{as_matrix}).

Precisely, a set of 42 sequences are returned, for a design with 14
periods and seven treatments. Ultimately, the
\ifelse{html}{\out{(<i>k</i>,<i>j</i>)}}{\eqn{(k,j)}}th
element of the cross-over design matrix corresponds to the treatment a
subject on the \ifelse{html}{\out{<i>k</i>}}{\eqn{k}}th sequence would
receive in the \ifelse{html}{\out{<i>j</i>}}{\eqn{j}}th period.
}
\examples{
# Anderson (2002) design for the default parameters
anderson         <- seq_anderson()
# Using different labels
anderson_ABCDEFG <- seq_anderson(labels = LETTERS[1:7])
}
\references{
Anderson I (2002) Training schedule design. \emph{Personal
Communication}.
}
\author{
Based on data from the \code{\link[Crossover]{Crossover}} package by
Kornelius Rohmeyer.
}
