% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_anderson_preece.R
\name{seq_anderson_preece}
\alias{seq_anderson_preece}
\title{Anderson and Preece (2002) cross-over design specification}
\usage{
seq_anderson_preece(labels = 0:6, as_matrix = T, summary = T)
}
\arguments{
\item{labels}{A \code{\link[base]{vector}} of labels for the treatments.
Should be of \code{\link[base]{length}} seven, containing unique elements.
Defaults to \code{0:6}.}

\item{as_matrix}{A \code{\link[base]{logical}} variable indicating whether
the design should be returned as a \code{\link[base]{matrix}}, or a
\code{\link[tibble]{tibble}}. Defaults to \code{T}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
Either a \code{\link[base]{matrix}} if \code{as_matrix = T} (with
rows corresponding to sequences and columns to periods), or a
\code{\link[tibble]{tibble}} if \code{as_matrix = F} (with rows corresponding
to a particular period on a particular sequence). In either case, the
returned object will have class \code{xover_seq}.
}
\description{
Specifies a cross-over design from Anderson and Preece (2002).
}
\details{
\code{seq_anderson_preece()} supports the specification of a cross-over
design from Anderson and Preece (2002). Sequences are supported for any
chosen treatment labels (see \code{labels}). In addition, the design can
be returned in \code{\link[base]{matrix}} or \code{\link[tibble]{tibble}}
form (see \code{as_matrix}).

Precisely, a set of 42 sequences are returned, for a design with seven
periods and seven treatments. Ultimately, the
\ifelse{html}{\out{(<i>k</i>,<i>j</i>)}}{\eqn{(k,j)}}th element of the
cross-over design matrix corresponds to the treatment a subject on the
\ifelse{html}{\out{<i>k</i>}}{\eqn{k}}th sequence would receive in the
\ifelse{html}{\out{<i>j</i>}}{\eqn{j}}th period.
}
\examples{
# Anderson and Preece (2002) design for the default parameters
anderson_preece         <- seq_anderson_preece()
# Using different labels
anderson_preece_ABCDEFG <- seq_anderson_preece(labels = LETTERS[1:7])
# Returning in tibble form
anderson_preece_tibble  <- seq_anderson_preece(as_matrix = F)
}
\references{
Anderson I, Preece DA (2002) Locally balanced change-over
designs. \emph{Utilitas Mathematica} \strong{62:}33-59.
}
\author{
Based on data from the \code{\link[Crossover]{Crossover}} package by
Kornelius Rohmeyer.
}
