% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_archdeacon.R
\name{seq_archdeacon}
\alias{seq_archdeacon}
\title{Archdeacon \emph{et al.} (1980) cross-over design specification}
\usage{
seq_archdeacon(labels = 0:8, as_matrix = T, summary = T)
}
\arguments{
\item{labels}{A \code{\link[base]{vector}} of labels for the treatments.
Should be of \code{\link[base]{length}} nine, containing unique elements.
Defaults to \code{0:8}.}

\item{as_matrix}{A \code{\link[base]{logical}} variable indicating whether
the design should be returned as a \code{\link[base]{matrix}}, or a
\code{\link[tibble]{tibble}}. Defaults to \code{T}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
Either a \code{\link[base]{matrix}} if \code{as_matrix = T} (with
rows corresponding to sequences and columns to periods), or a
\code{\link[tibble]{tibble}} if \code{as_matrix = F} (with rows corresponding
to a particular period on a particular sequence). In either case, the
returned object will have class \code{xover_seq}.
}
\description{
Specifies a cross-over design from Archdeacon \emph{et al.} (1980).
}
\details{
\code{seq_archdeacon()} supports the specification of cross-over designs from
Archdeacon \emph{et al.} (1980). Sequences are supported for any chosen
treatment labels (see \code{labels}). In addition, the design can be
returned in \code{\link[base]{matrix}} or \code{\link[tibble]{tibble}} form
(see \code{as_matrix}).

Precisely, a set of nine sequences are returned, for a design with nine
periods and nine treatments. Ultimately, the
\ifelse{html}{\out{(<i>k</i>,<i>j</i>)}}{\eqn{(k,j)}}th
element of the cross-over design matrix corresponds to the treatment a
subject on the \ifelse{html}{\out{<i>k</i>}}{\eqn{k}}th sequence would
receive in the \ifelse{html}{\out{<i>j</i>}}{\eqn{j}}th period.
}
\examples{
# Archdeacon et al. (1980) design for the default parameters
archdeacon         <- seq_archdeacon()
# Using different labels
archdeacon_LETTERS <- seq_archdeacon(labels = LETTERS[1:9])
# Returning in tibble form
archdeacon_tibble  <- seq_archdeacon(as_matrix = F)
}
\references{
Archdeacon DS, Dinitz JH, Stinson DR, Tillson TW (1980) Some new
row-complete latin squares. \emph{J Comb Theory A} \strong{29:}395-398.
}
\author{
Based on data from the \code{\link[Crossover]{Crossover}} package by
Kornelius Rohmeyer.
}
