% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_balmin_rmd.R
\name{seq_balmin_rmd}
\alias{seq_balmin_rmd}
\title{Afsarinejad (1983) cross-over design specification}
\usage{
seq_balmin_rmd(D = 2, J = 2, labels = 0:(D - 1), as_matrix = T,
  summary = T)
}
\arguments{
\item{D}{The number of treatments. Must be a single
\code{\link[base]{numeric}} integer greater than or equal to two. Defaults to
\code{2}.}

\item{J}{The number of periods. Must be a single \code{\link[base]{numeric}}
integer less than or equal to \code{D}. Defaults to \code{2}.}

\item{labels}{A \code{\link[base]{vector}} of labels for the treatments.
Should be of \code{\link[base]{length}} \code{D}, containing unique elements.
Defaults to \code{0:(D - 1)}.}

\item{as_matrix}{A \code{\link[base]{logical}} variable indicating whether
the design should be returned as a \code{\link[base]{matrix}}, or a
\code{\link[tibble]{tibble}}. Defaults to \code{T}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
Either a \code{\link[base]{matrix}} if \code{as_matrix = T} (with
rows corresponding to sequences and columns to periods), or a
\code{\link[tibble]{tibble}} if \code{as_matrix = F} (with rows corresponding
to a particular period on a particular sequence). In either case, the
returned object will have class \code{xover_seq}.
}
\description{
Specifies cross-over designs from Afsarinejad (1983).
}
\details{
\code{seq_balmin_rmd()} supports the specification of cross-over designs from
Afsarinejad (1983). Designs for any number of treatments (see \code{D}) are
supported, for any chosen treatment labels (see \code{labels}). The number of
periods (see \code{J}) can be any number such that
\ifelse{html}{\out{(<i>D</i> - 1)/(<i>J</i> - 1)}}{\eqn{(D - 1)/(J - 1)}} is
a positive integer. In addition, the designs can be returned in
\code{\link[base]{matrix}} or \code{\link[tibble]{tibble}} form (see
\code{as_matrix}).

Precisely, a balanced minimal repeated measurement design is returned. The
sequences are incomplete, i.e., no sequence contains all of the treatments.
The design is balanced for carryover effects but will in general not be a
balanced block design. A necessary and sufficient condition for the existence
of such a design is that
\ifelse{html}{\out{(<i>D</i> - 1)/(<i>J</i> - 1)}}{\eqn{(D - 1)/(J - 1)}} is
a positive integer: this is therefore enforced by the function. The number of
sequences, \ifelse{html}{\out{<i>K</i>}}{\eqn{K}}, in the resulting design is
\ifelse{html}{\out{<i>D</i>(<i>D</i> - 1)/(<i>J</i> - 1)}}{\eqn{
D(D - 1)/(J - 1)}}. Ultimately, the
\ifelse{html}{\out{(<i>k</i>,<i>j</i>)}}{\eqn{(k,j)}}th element of the
cross-over design matrix corresponds to the treatment a subject on the
\ifelse{html}{\out{<i>k</i>}}{\eqn{k}}th sequence would receive in the
\ifelse{html}{\out{<i>j</i>}}{\eqn{j}}th period.
}
\examples{
# Afsarinejad (1983) design for four treatments and two periods
balmin_rmd        <- seq_balmin_rmd()
# Using different labels
balmin_rmd_ABCD   <- seq_balmin_rmd(labels = LETTERS[1:4])
# Returning in tibble form
balmin_rmd_tibble <- seq_balmin_rmd(as_matrix = F)
}
\references{
Afsarinejad K (1983) Balanced repeated measurements designs.
\emph{Biometrika} \strong{70:}199-204.

Wakeling IN, MacFie HJH (1995) Designing consumer trials balanced
for first and higher orders of carry-over effect when only a subset of k
samples from t may be tested. \emph{Food Qual Prefer} \strong{6:}299-308.
}
\author{
Based on code from the \code{\link[crossdes]{crossdes}} package by
Oliver Sailer.
}
