% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_bate_jones.R
\name{seq_bate_jones}
\alias{seq_bate_jones}
\title{Bate and Jones (2002) cross-over design specification}
\usage{
seq_bate_jones(D = 5, labels = 0:(D - 1), as_matrix = T,
  summary = T)
}
\arguments{
\item{D}{The number of treatments. Must be either five or eight. Defaults to
\code{2}.}

\item{labels}{A \code{\link[base]{vector}} of labels for the treatments.
Should be of \code{\link[base]{length}} \code{D}, containing unique elements.
Defaults to \code{0:(D - 1)}.}

\item{as_matrix}{A \code{\link[base]{logical}} variable indicating whether
the design should be returned as a \code{\link[base]{matrix}}, or a
\code{\link[tibble]{tibble}}. Defaults to \code{T}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
Either a \code{\link[base]{matrix}} if \code{as_matrix = T} (with
rows corresponding to sequences and columns to periods), or a
\code{\link[tibble]{tibble}} if \code{as_matrix = F} (with rows corresponding
to a particular period on a particular sequence). In either case, the
returned object will have class \code{xover_seq}.
}
\description{
Specifies cross-over designs from Bate and Jones (2002).
}
\details{
\code{seq_bate_jones()} supports the specification of cross-over designs from
Bate and Jones (2002). Designs for five and eight treatments (see \code{D})
are supported, for any chosen treatment labels (see \code{labels}). In
addition, the designs can be returned in \code{\link[base]{matrix}} or
\code{\link[tibble]{tibble}} form (see \code{as_matrix}).

Precisely, the \ifelse{html}{\out{(<i>k</i>,<i>j</i>)}}{\eqn{(k,j)}}th
element of the cross-over design matrix corresponds to the treatment a
subject on the \ifelse{html}{\out{<i>k</i>}}{\eqn{k}}th sequence would
receive in the \ifelse{html}{\out{<i>j</i>}}{\eqn{j}}th period.
}
\examples{
# Bate and Jones (2002) design for five treatments
bate_jones        <- seq_bate_jones()
# Using different labels
bate_jones_ABCDE  <- seq_bate_jones(labels = LETTERS[1:5])
# Returning in tibble form
bate_jones_tibble <- seq_bate_jones(as_matrix = F)
}
\references{
Bate S, Jones B (2002) The construction of universally optimal
uniform cross-over designs. \emph{GlaxoSmithKline Biomedical Data Sciences
Technical Report}.
}
\author{
Based on data from the \code{\link[Crossover]{Crossover}} package by
Kornelius Rohmeyer.
}
