% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_clatworthy.R
\name{seq_clatworthy}
\alias{seq_clatworthy}
\title{Clatworthy (1973) cross-over design specification}
\usage{
seq_clatworthy(type = "R", selection = 1, labels, as_matrix = T,
  summary = T)
}
\arguments{
\item{type}{A \code{\link[base]{character}} string indicating which type of
sequences to return. Must be one of \code{"R"}, \code{"S"}, and \code{"SR"}.
Defaults to \code{"R"}.}

\item{selection}{A \code{\link[base]{numeric}} integer or
\code{\link[base]{character}} string indicating which design to return, for
the chosen value of \code{type}. See \strong{Details} for information on how
to ascertain supported values. Defaults to \code{1}.}

\item{labels}{A \code{\link[base]{vector}} of labels for the treatments.
If specified, should have \code{\link[base]{length}} corresponding to the
number of treatments in the chosen design, containing unique elements. If
missing, will internally default to \code{0:(D - 1)}, where \code{D} is the
number of treatments.}

\item{as_matrix}{A \code{\link[base]{logical}} variable indicating whether
the design should be returned as a \code{\link[base]{matrix}}, or a
\code{\link[tibble]{tibble}}. Defaults to \code{T}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
Either a \code{\link[base]{matrix}} if \code{as_matrix = T} (with
rows corresponding to sequences and columns to periods), or a
\code{\link[tibble]{tibble}} if \code{as_matrix = F} (with rows corresponding
to a particular period on a particular sequence). In either case, the
returned object will have class \code{xover_seq}.
}
\description{
Specifies cross-over designs from Clatworthy (1973).
}
\details{
\code{seq_clatworthy()} supports the specification of designs from
Clatworthy (1973). Designs for a large array of values of the number of
treatments and periods are available (accessible by setting \code{selection}
and \code{type} carefully), for any chosen treatment labels (see
\code{labels}). In addition, the designs can be returned in
\code{\link[base]{matrix}} or \code{\link[tibble]{tibble}} form (see
\code{as_matrix}).

Precisely, \code{type} must be set as one of \code{"R"}, \code{"S"}, and
\code{"SR"}. Then, \code{selection} can take one of a particular set of
\code{\link[base]{numeric}} or \code{\link[base]{character}} values. The best
way to ascertain which values are allowed is to use the utility function
\code{\link[xover]{summary_seq_clatworthy}}. Ultimately, the
\ifelse{html}{\out{(<i>k</i>,<i>j</i>)}}{\eqn{(k,j)}}th element of the
cross-over design matrix corresponds to the treatment a subject on the
\ifelse{html}{\out{<i>k</i>}}{\eqn{k}}th sequence would receive in the
\ifelse{html}{\out{<i>j</i>}}{\eqn{j}}th period.
}
\examples{
# A selection of designs for four treatments from Clatworthy (1973)
clatworthy1        <- seq_clatworthy()
clatworthy2        <- seq_clatworthy(selection = 2)
clatworthy3        <- seq_clatworthy(selection = 3)
clatworthy4        <- seq_clatworthy(selection = 4)
# Using different labels
clatworthy1_ABCD   <- seq_clatworthy(labels = LETTERS[1:4])
clatworthy2_ABCD   <- seq_clatworthy(selection = 2, labels = LETTERS[1:4])
clatworthy3_ABCD   <- seq_clatworthy(selection = 3, labels = LETTERS[1:4])
clatworthy4_ABCD   <- seq_clatworthy(selection = 4, labels = LETTERS[1:4])
# Returning in tibble form
clatworthy1_tibble <- seq_clatworthy(as_matrix = F)
clatworthy2_tibble <- seq_clatworthy(selection = 2, as_matrix = F)
clatworthy3_tibble <- seq_clatworthy(selection = 3, as_matrix = F)
clatworthy4_tibble <- seq_clatworthy(selection = 4, as_matrix = F)
}
\references{
Clatworthy WH, Cameron JM, Speckman JA (1973) Tables of
two-associate-class partially balanced designs. \emph{US Government Printing
Office}.
}
\author{
Based on data from the \code{\link[Crossover]{Crossover}} package by
Kornelius Rohmeyer.
}
