% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_davis_hall.R
\name{seq_davis_hall}
\alias{seq_davis_hall}
\title{Davis and Hall (1969) cross-over design specification}
\usage{
seq_davis_hall(D = 6, selection = 1, labels = 0:(D - 1),
  as_matrix = T, summary = T)
}
\arguments{
\item{D}{The number of treatments. Must be either six, seven, eight, or nine.
Defaults to \code{6}.}

\item{selection}{A single \code{\link[base]{numeric}} integer indicating
which design to return, for the chosen value of \code{D}. See
\strong{Details} for information on supported values.}

\item{labels}{A \code{\link[base]{vector}} of labels for the treatments.
Should be of \code{\link[base]{length}} \code{D}, containing unique elements.
Defaults to \code{0:(D - 1)}.}

\item{as_matrix}{A \code{\link[base]{logical}} variable indicating whether
the design should be returned as a \code{\link[base]{matrix}}, or a
\code{\link[tibble]{tibble}}. Defaults to \code{T}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
Either a \code{\link[base]{matrix}} if \code{as_matrix = T} (with
rows corresponding to sequences and columns to periods), or a
\code{\link[tibble]{tibble}} if \code{as_matrix = F} (with rows corresponding
to a particular period on a particular sequence). In either case, the
returned object will have class \code{xover_seq}.
}
\description{
Specifies cross-over designs from Davis and Hall (1969).
}
\details{
\code{seq_davis_hall()} supports the specification of designs from
Davis and Hall (1969). Designs for six through nine treatments (see
\code{D}) are supported, for any chosen treatment labels (see \code{labels}).
In addition, the designs can be returned in \code{\link[base]{matrix}} or
\code{\link[tibble]{tibble}} form (see \code{as_matrix}).

Precisely, for each supported value of \code{D}, there are three designs
available (accessible by setting \code{selection} equal to one, two, or
three). Ultimately, the
\ifelse{html}{\out{(<i>k</i>,<i>j</i>)}}{\eqn{(k,j)}}th element of the
cross-over design matrix corresponds to the treatment a subject on the
\ifelse{html}{\out{<i>k</i>}}{\eqn{k}}th sequence would receive in the
\ifelse{html}{\out{<i>j</i>}}{\eqn{j}}th period.
}
\examples{
# Davis and Hall (1969) designs for six treatments
davis_hall1        <- seq_davis_hall()
davis_hall2        <- seq_davis_hall(selection = 2)
davis_hall3        <- seq_davis_hall(selection = 3)
# Using different labels
davis_hall1_ABCDEF <- seq_davis_hall(labels = LETTERS[1:6])
davis_hall2_ABCDEF <- seq_davis_hall(selection = 2, labels = LETTERS[1:6])
davis_hall3_ABCDEF <- seq_davis_hall(selection = 3, labels = LETTERS[1:6])
# Returning in tibble form
davis_hall1_tibble <- seq_davis_hall(as_matrix = F)
davis_hall2_tibble <- seq_davis_hall(selection = 2, as_matrix = F)
davis_hall3_tibble <- seq_davis_hall(selection = 3, as_matrix = F)
}
\references{
Davis AW, Hall WB (1969) Cyclic change-over designs.
\emph{Biometrika} \strong{56:}283-93.
}
\author{
Based on data from the \code{\link[Crossover]{Crossover}} package by
Kornelius Rohmeyer.
}
