% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_fletcher.R
\name{seq_fletcher}
\alias{seq_fletcher}
\title{Fletcher (1987) cross-over design specification}
\usage{
seq_fletcher(D = 6, selection = 1, labels = 0:(D - 1),
  as_matrix = T, summary = T)
}
\arguments{
\item{D}{The number of treatments. Must be either four, six, eight, or nine.
Defaults to \code{4}.}

\item{selection}{A single \code{\link[base]{numeric}} integer indicating
which design to return, for the chosen value of \code{D}. See
\strong{Details} for information on supported values.}

\item{labels}{A \code{\link[base]{vector}} of labels for the treatments.
Should be of \code{\link[base]{length}} \code{D}, containing unique elements.
Defaults to \code{0:(D - 1)}.}

\item{as_matrix}{A \code{\link[base]{logical}} variable indicating whether
the design should be returned as a \code{\link[base]{matrix}}, or a
\code{\link[tibble]{tibble}}. Defaults to \code{T}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
Either a \code{\link[base]{matrix}} if \code{as_matrix = T} (with
rows corresponding to sequences and columns to periods), or a
\code{\link[tibble]{tibble}} if \code{as_matrix = F} (with rows corresponding
to a particular period on a particular sequence). In either case, the
returned object will have class \code{xover_seq}.
}
\description{
Specifies cross-over designs from Fletcher (1987).
}
\details{
\code{seq_fletcher()} supports the specification of designs from Fletcher
(1987). Deisgns for four, six, eight, and nine treatments (see
\code{D}) are supported, for any chosen treatment labels (see \code{labels}).
In addition, the designs can be returned in \code{\link[base]{matrix}} or
\code{\link[tibble]{tibble}} form (see \code{as_matrix}).

Precisely, for \code{D} equal to four, six, eight, and nine, there are four,
three, 12, and nine designs available respectively (accessible by setting
\code{selection} equal to one through twelve asappropriate). Ultimately, the
\ifelse{html}{\out{(<i>k</i>,<i>j</i>)}}{\eqn{(k,j)}}th element of the
cross-over design matrix corresponds to the treatment a subject on the
\ifelse{html}{\out{<i>k</i>}}{\eqn{k}}th sequence would receive in the
\ifelse{html}{\out{<i>j</i>}}{\eqn{j}}th period.
}
\examples{
# Fletcher (1987) designs for four treatments
fletcher1        <- seq_fletcher(D = 4)
fletcher2        <- seq_fletcher(D = 4, selection = 2)
fletcher3        <- seq_fletcher(D = 4, selection = 3)
fletcher4        <- seq_fletcher(D = 4, selection = 4)
# Using different labels
fletcher1_ABCD   <- seq_fletcher(D = 4, labels = LETTERS[1:4])
fletcher2_ABCD   <- seq_fletcher(D = 4, selection = 2, labels = LETTERS[1:4])
fletcher3_ABCD   <- seq_fletcher(D = 4, selection = 3, labels = LETTERS[1:4])
fletcher4_ABCD   <- seq_fletcher(D = 4, selection = 4, labels = LETTERS[1:4])
# Returning in tibble form
fletcher1_tibble <- seq_fletcher(D = 4, as_matrix = F)
fletcher2_tibble <- seq_fletcher(D = 4, selection = 2, as_matrix = F)
fletcher3_tibble <- seq_fletcher(D = 4, selection = 3, as_matrix = F)
fletcher4_tibble <- seq_fletcher(D = 4, selection = 4, as_matrix = F)
}
\references{
Fletcher DJ (1987) A new class of change-over designs for
factorial experiments. \emph{Biometrika} \strong{74:}649-54.
}
\author{
Based on data from the \code{\link[Crossover]{Crossover}} package by
Kornelius Rohmeyer.
}
