% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_iqbal_jones.R
\name{seq_iqbal_jones}
\alias{seq_iqbal_jones}
\title{Iqbal and Jones (1994) cross-over design specification}
\usage{
seq_iqbal_jones(D = 3, selection = 1, labels = 0:(D - 1),
  as_matrix = T, summary = T)
}
\arguments{
\item{D}{The number of treatments. Must be an integer between three and ten
inclusive. Defaults to \code{3}.}

\item{selection}{A single \code{\link[base]{numeric}} integer indicating
which design to return, for the chosen value of \code{D}. See
\strong{Details} for information on supported values.}

\item{labels}{A \code{\link[base]{vector}} of labels for the treatments.
Should be of \code{\link[base]{length}} \code{D}, containing unique elements.
Defaults to \code{0:(D - 1)}.}

\item{as_matrix}{A \code{\link[base]{logical}} variable indicating whether
the design should be returned as a \code{\link[base]{matrix}}, or a
\code{\link[tibble]{tibble}}. Defaults to \code{T}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
Either a \code{\link[base]{matrix}} if \code{as_matrix = T} (with
rows corresponding to sequences and columns to periods), or a
\code{\link[tibble]{tibble}} if \code{as_matrix = F} (with rows corresponding
to a particular period on a particular sequence). In either case, the
returned object will have class \code{xover_seq}.
}
\description{
Specifies cross-over designs from Iqbal and Jones (1994).
}
\details{
\code{seq_iqbal_jones()} supports the specification of designs from
Iqbal and Jones (1994). Designs for three through ten treatments (see
\code{D}) are supported, for any chosen treatment labels (see \code{labels}).
In addition, the designs can be returned in \code{\link[base]{matrix}} or
\code{\link[tibble]{tibble}} form (see \code{as_matrix}).

Precisely, for \code{D} equal to three through ten, there are five, three,
four, five, six, six, one, and one designs available respectively (accessible
by setting \code{selection} equal to one through six as appropriate).
Ultimately, the
\ifelse{html}{\out{(<i>k</i>,<i>j</i>)}}{\eqn{(k,j)}}th element of the
cross-over design matrix corresponds to the treatment a subject on the
\ifelse{html}{\out{<i>k</i>}}{\eqn{k}}th sequence would receive in the
\ifelse{html}{\out{<i>j</i>}}{\eqn{j}}th period.
}
\examples{
# Iqbal and Jones (1994) designs for four treatments
iqbal_jones1        <- seq_iqbal_jones(D = 4)
iqbal_jones2        <- seq_iqbal_jones(D = 4, selection = 2)
iqbal_jones3        <- seq_iqbal_jones(D = 4, selection = 3)
# Using different labels
iqbal_jones1_ABCD   <- seq_iqbal_jones(D = 4, labels = LETTERS[1:4])
iqbal_jones2_ABCD   <- seq_iqbal_jones(D = 4, selection = 2,
                                       labels = LETTERS[1:4])
iqbal_jones3_ABCD   <- seq_iqbal_jones(D = 4, selection = 3,
                                       labels = LETTERS[1:4])
# Returning in tibble form
iqbal_jones1_tibble <- seq_iqbal_jones(D = 4, as_matrix = F)
iqbal_jones2_tibble <- seq_iqbal_jones(D = 4, selection = 2, as_matrix = F)
iqbal_jones3_tibble <- seq_iqbal_jones(D = 4, selection = 3, as_matrix = F)
}
\references{
Iqbal I, Jones B (1994) Efficient repeated measurements designs
with equal and unequal period sizes. \emph{J Stat Plan Infer}
\strong{42:}79-88.
}
\author{
Based on data from the \code{\link[Crossover]{Crossover}} package by
Kornelius Rohmeyer.
}
