% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_lewis.R
\name{seq_lewis}
\alias{seq_lewis}
\title{Lewis \emph{et al.} (1988) cross-over design specification}
\usage{
seq_lewis(D = 4, selection = 1, labels = 0:(D - 1), as_matrix = T,
  summary = T)
}
\arguments{
\item{D}{The number of treatments. Must be either four, six, eight, or nine.
Defaults to \code{4}.}

\item{selection}{A single \code{\link[base]{numeric}} integer indicating
which design to return, for the chosen value of \code{D}. See
\strong{Details} for information on supported values.}

\item{labels}{A \code{\link[base]{vector}} of labels for the treatments.
Should be of \code{\link[base]{length}} \code{D}, containing unique elements.
Defaults to \code{0:(D - 1)}.}

\item{as_matrix}{A \code{\link[base]{logical}} variable indicating whether
the design should be returned as a \code{\link[base]{matrix}}, or a
\code{\link[tibble]{tibble}}. Defaults to \code{T}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
Either a \code{\link[base]{matrix}} if \code{as_matrix = T} (with
rows corresponding to sequences and columns to periods), or a
\code{\link[tibble]{tibble}} if \code{as_matrix = F} (with rows corresponding
to a particular period on a particular sequence). In either case, the
returned object will have class \code{xover_seq}.
}
\description{
Specifies cross-over designs from Lewis \emph{et al.} (1988).
}
\details{
\code{seq_lewis()} supports the specification of designs from Lewis
\emph{et al.} (1988). Designs for four, six, eight, and nine treatments (see
\code{D}) are supported, for any chosen treatment labels (see \code{labels}).
In addition, the designs can be returned in \code{\link[base]{matrix}} or
\code{\link[tibble]{tibble}} form (see \code{as_matrix}).

Precisely, for \code{D} equal to four, six, eight and nine, there are six,
eight, four, and two sets of sequences available respectively (accessible by
setting \code{selection} equal to one through eight as appropriate).
Ultimately, the \ifelse{html}{\out{(<i>k</i>,<i>j</i>)}}{\eqn{(k,j)}}th
element of the cross-over design matrix corresponds to the treatment a
subject on the \ifelse{html}{\out{<i>k</i>}}{\eqn{k}}th sequence would
receive in the \ifelse{html}{\out{<i>j</i>}}{\eqn{j}}th period.
}
\examples{
# Lewis et al. (1988) designs for nine treatments
lewis1         <- seq_lewis(D = 9)
lewis2         <- seq_lewis(D = 9, selection = 2)
# Using different labels
lewis1_LETTERS <- seq_lewis(D = 9, labels = LETTERS[1:9])
lewis2_LETTERS <- seq_lewis(D = 9, selection = 2, labels = LETTERS[1:9])
# Returning in tibble form
lewis1_tibble  <- seq_lewis(D = 9, as_matrix = F)
lewis2_tibble  <- seq_lewis(D = 9, selection = 2, as_matrix = F)
}
\references{
Lewis SM, Fletcher DJ, Matthews JNS (1988) Factorial cross-over
designs in clinical trials. In \emph{Optimal Design and Analysis of
Experiments}, Eds: Dodge Y, Fedorov VV, Wynn HP, pp 133-140. Elsevier Science
Publishers BV: North-Holland.
}
\author{
Based on data from the \code{\link[Crossover]{Crossover}} package by
Kornelius Rohmeyer.
}
