% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_mols.R
\name{seq_mols}
\alias{seq_mols}
\title{Mutually Orthogonal Latin Square cross-over design specification}
\usage{
seq_mols(D = 4, type = "set", J = D, labels = 0:(D - 1),
  as_matrix = T, summary = T)
}
\arguments{
\item{D}{The number of treatments. Must be a single
\code{\link[base]{numeric}} prime power between three and 100. Defaults to
\code{3}.}

\item{type}{Must be either \code{"set"} or \code{"single"}, indicating
whether a complete set of MOLS, or a single cross-over design based on such a
complete set, should be returned. Defaults to \code{"set"}.}

\item{J}{When \code{type = "single"}, this will determine
the number of periods in the resulting design. Must be a single
\code{\link[base]{numeric}} integer between two and \code{D} inclusive.
Defaults to \code{D}.}

\item{labels}{A \code{\link[base]{vector}} of labels for the treatments.
Should be of \code{\link[base]{length}} \code{D}, containing unique elements.
Defaults to \code{0:(D - 1)}.}

\item{as_matrix}{A \code{\link[base]{logical}} variable indicating whether
the design should be returned as a \code{\link[base]{matrix}}, or a
\code{\link[tibble]{tibble}}. Defaults to \code{T}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
If \code{type = "single"}: Either a \code{\link[base]{matrix}} if
\code{as_matrix = T} (with rows corresponding to sequences and columns to
periods), or a \code{\link[tibble]{tibble}} if \code{as_matrix = F} (with
rows corresponding to a particular period on a particular sequence).
Otherwise, if \code{type = "set"}, a list consisting of either matrices or
tibbles, again dependent on the value of \code{as_matrix}. In either case,
the returned object(s) will have class \code{xover_seq}.
}
\description{
Specifies either a complete set of Mutually Orthogonal Latin Squares (MOLS),
or a single cross-over design based on a complete set of MOLS.
}
\details{
\code{seq_mols()} supports the specification of either a complete set of
MOLS, or a single cross-over design based on such a set (see \code{type}).
Designs for any prime power number of treatments (see \code{D}) are
supported, for any chosen treatment labels (see \code{labels}). In
addition, the designs can be returned in \code{\link[base]{matrix}} or
\code{\link[tibble]{tibble}} form (see \code{as_matrix}). When a single
cross-over design is to be returned, the number of periods can be any integer
between two and \code{D}. Ultimately,
the \ifelse{html}{\out{(<i>k</i>,<i>j</i>)}}{\eqn{(k,j)}}th
element of the cross-over design matrix/matrices corresponds to the treatment
a subject on the \ifelse{html}{\out{<i>k</i>}}{\eqn{k}}th sequence would
receive in the \ifelse{html}{\out{<i>j</i>}}{\eqn{j}}th period.

The complete set of MOLS is constructed using Galois fields. In the case that
\code{type = "set"}, \ifelse{html}{\out{<i>D</i> - 1}}{\eqn{D - 1}} Latin
Squares of order \ifelse{html}{\out{<i>D</i>}}{\eqn{D}} will be constructed.
They will be mutually orthogonal, i.e., if any two of them are super-imposed,
the resulting array will contain each ordered pair
\ifelse{html}{\out{(<i>d</i><sub>1</sub>,<i>d</i><sub>2
</sub>)}}{\eqn{(d_1,d_2)}} exactly once. The Latin Squares will be in
standard order, i.e., the first sequence is always equal to the specified
treatment labels. Otherwise, when \code{type = "single"}, the resulting
design will have \ifelse{html}{\out{<i>D</i>(<i>D</i> - 1)}}{\eqn{D(D - 1)}}
sequences, and each treatment will appear in each sequence at most once. The
design will be a generalized Youden design that is also balanced for
carryover effects.
}
\examples{
# A complete set of MOLS for three treatments
mols_set    <- seq_mols()
# A corresponding cross-over design, based on a complete set of MOLS
mols_single <- seq_mols(type = "single")
}
\references{
Wakeling IN, MacFie HJH (1995) Designing consumer trials balanced
for first and higher orders of carry-over effect when only a subset of k
samples from t may be tested. \emph{Food Qual Prefer} \strong{6:}299-308.

Williams EJ (1949) Experimental designs balanced for the
estimation of residual effects of treatments. \emph{Aust J Sci Res Ser A}
\strong{2:}149-168.
}
\author{
Based on functions from the \code{\link[crossdes]{crossdes}} package
by Oliver Sailer.
}
