% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_patterson_lucas.R
\name{seq_patterson_lucas}
\alias{seq_patterson_lucas}
\title{Patterson and Lucas (1962) cross-over design specification}
\usage{
seq_patterson_lucas(selection = 1, labels, as_matrix = T,
  summary = T)
}
\arguments{
\item{selection}{A single \code{\link[base]{numeric}} integer indicating
which design to return. See \strong{Details} for information on supported
values.}

\item{labels}{A \code{\link[base]{vector}} of labels for the treatments.
If specified, should have \code{\link[base]{length}} corresponding to the
number of treatments in the chosen design, containing unique elements. If
missing, will internally default to \code{0:(D - 1)}, where \code{D} is the
number of treatments.}

\item{as_matrix}{A \code{\link[base]{logical}} variable indicating whether
the design should be returned as a \code{\link[base]{matrix}}, or a
\code{\link[tibble]{tibble}}. Defaults to \code{T}.}
}
\value{
Either a \code{\link[base]{matrix}} if \code{as_matrix = T} (with
rows corresponding to sequences and columns to periods), or a
\code{\link[tibble]{tibble}} if \code{as_matrix = F} (with rows corresponding
to a particular period on a particular sequence). In either case, the
returned object will have class \code{xover_seq}.
}
\description{
Specifies cross-over designs from Patterson and Lucas (1962).
}
\details{
\code{seq_patterson_lucas()} supports the specification of designs from
Patterson and Lucas (1962). Designs for three through nine treatments (see
\code{D}) are supported, for any chosen treatment labels (see \code{labels}).
In addition, the designs can be returned in \code{\link[base]{matrix}} or
\code{\link[tibble]{tibble}} form (see \code{as_matrix}).

Precisely, designs are indexed using the argument \code{selection}, which can
take the following values for the following numbers of treatments:

\itemize{
\item \code{D = 3}: 1, 30, and 31;
\item \code{D = 4}: 3-5, 32-35, and 153;
\item \code{D = 5}: 7-10, 18, 36-40, 125, 126, 131, 132, and 154;
\item \code{D = 6}: 12, 13, 15, 41, 42, 99-101, 133-135, and 155;
\item \code{D = 7}: 16, 17, 19-23, 43-49, and 86;
\item \code{D = 8}: 102-105, 136-138, and 156;
\item \code{D = 9}: 106, 107, 127, 128, and 139-141.
}

Ultimately, the \ifelse{html}{\out{(<i>k</i>,<i>j</i>)}}{\eqn{(k,j)}}th
element of the cross-over design matrix corresponds to the treatment a
subject on the \ifelse{html}{\out{<i>k</i>}}{\eqn{k}}th sequence would
receive in the \ifelse{html}{\out{<i>j</i>}}{\eqn{j}}th period.
}
\examples{
# Patterson and Lucas (1962) designs for three treatments
patterson_lucas1         <- seq_patterson_lucas()
patterson_lucas30        <- seq_patterson_lucas(selection = 30)
patterson_lucas31        <- seq_patterson_lucas(selection = 31)
# Using different labels
patterson_lucas1_ABC     <- seq_patterson_lucas(labels = LETTERS[1:3])
patterson_lucas30_ABC    <- seq_patterson_lucas(selection = 30,
                                                labels = LETTERS[1:3])
patterson_lucas31_ABC    <- seq_patterson_lucas(selection = 31,
                                                 labels = LETTERS[1:3])
# Returning in tibble form
patterson_lucas1_tibble  <- seq_patterson_lucas(as_matrix = F)
patterson_lucas30_tibble <- seq_patterson_lucas(selection = 30,
as_matrix = F)
patterson_lucas31_tibble <- seq_patterson_lucas(selection = 31,
                                                as_matrix = F)
}
\references{
Patterson HD, Lucas HL (1962) Change-over designs. \emph{North
Carolina Agricultural Experiment Station}. Tech Bull 147.
}
\author{
Based on data from the \code{\link[Crossover]{Crossover}} package by
Kornelius Rohmeyer.
}
