% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_clatworthy_williams.R
\name{seq_pbib_clatworthy_williams}
\alias{seq_pbib_clatworthy_williams}
\title{Clatworthy-Williams cross-over design specification}
\usage{
seq_pbib_clatworthy_williams(williams_D = 3, type = "R",
  selection = 42, labels, as_matrix = T, summary = T)
}
\arguments{
\item{type}{A \code{\link[base]{character}} string indicating which type of
sequences to return. Must be one of \code{"R"}, \code{"S"}, and \code{"SR"}.
Defaults to \code{"R"}.}

\item{selection}{A \code{\link[base]{numeric}} integer or
\code{\link[base]{character}} string indicating which design to return, for
the chosen values of \code{williams_D} and \code{type}. See \strong{Details}
for information on how to ascertain supported values. Defaults to \code{1}.}

\item{labels}{A \code{\link[base]{vector}} of labels for the treatments.
If specified, should have \code{\link[base]{length}} corresponding to the
number of treatments in the chosen design, containing unique elements. If
missing, will internally default to \code{0:(D - 1)}, where \code{D} is the
number of treatments.}

\item{as_matrix}{A \code{\link[base]{logical}} variable indicating whether
the design should be returned as a \code{\link[base]{matrix}}, or a
\code{\link[tibble]{tibble}}. Defaults to \code{T}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}

\item{D}{The number of treatments in the utilised Williams (1949) design.
Must be a single \code{\link[base]{numeric}} integer between three and nine
inclusive. Defaults to \code{3}.}
}
\value{
Either a \code{\link[base]{matrix}} if \code{as_matrix = T} (with
rows corresponding to sequences and columns to periods), or a
\code{\link[tibble]{tibble}} if \code{as_matrix = F} (with rows corresponding
to a particular period on a particular sequence). In either case, the
returned object will have class \code{xover_seq}.
}
\description{
Specifies cross-over designs based on combining Clatworthy (1973) designs
with Williams (1949) designs.
}
\details{
\code{seq_clatworthy_williams()} supports the specification of designs based
on combining designs from Clatworthy (1973) with designs from Williams
(1949). Designs for a large array of values of the number of
treatments and periods are available (accessible by setting
\code{williams_D}, \code{selection} and \code{type} carefully), for any
chosen treatment labels (see \code{labels}). In addition, the designs can be
returned in \code{\link[base]{matrix}} or \code{\link[tibble]{tibble}} form
(see \code{as_matrix}).

Precisely, \code{type} must be set as one of \code{"R"}, \code{"S"}, and
\code{"SR"}. Then, \code{williams_D} must be set to an integer between three
and nine inclusive. With this, \code{selection} can take one of a particular
set of \code{\link[base]{numeric}} or \code{\link[base]{character}} values.
The best way to ascertain which values are allowed is to use the utility
function \code{\link[xover]{summary_seq_clatworthy_williams}}. Ultimately,
the \ifelse{html}{\out{(<i>k</i>,<i>j</i>)}}{\eqn{(k,j)}}th element of the
cross-over design matrix corresponds to the treatment a subject on the
\ifelse{html}{\out{<i>k</i>}}{\eqn{k}}th sequence would receive in the
\ifelse{html}{\out{<i>j</i>}}{\eqn{j}}th period.
}
\examples{
# A six treatment design
clatworthy_williams        <- seq_clatworthy_williams()
# Using different labels
clatworthy_williams_ABCDEF <- seq_clatworthy(labels = LETTERS[1:6])
# Returning in tibble form
clatworthy_williams_tibble <- seq_clatworthy(as_matrix = F)
}
\references{
Clatworthy WH, Cameron JM, Speckman JA (1973) Tables of
two-associate-class partially balanced designs. \emph{US Government Printing
Office}.

Patterson HD, Lucas HL (1962) Change-over designs. \emph{North
Carolina Agricultural Experiment Station}. Tech Bull 147.

Williams EJ (1949) Experimental designs balanced for the
estimation of residual effects of treatments. \emph{Aust J Sci Res Ser A}
\strong{2:}149-168.
}
\seealso{
\code{\link[xover]{summary_seq_clatworthy_williams}}.
}
\author{
Based on data from the \code{\link[Crossover]{Crossover}} package by
Kornelius Rohmeyer.
}
