% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_pigeon.R
\name{seq_pigeon}
\alias{seq_pigeon}
\title{Pigeon (1984) cross-over design specification}
\usage{
seq_pigeon(D = 3, selection = 1, labels = 0:(D - 1), as_matrix = T,
  summary = T)
}
\arguments{
\item{D}{The number of treatments. Must be a single
\code{\link[base]{numeric}} integer between three and seven inclusive.
Defaults to \code{3}.}

\item{selection}{A single \code{\link[base]{numeric}} integer indicating
which design to return, for the chosen value of \code{D}. See
\strong{Details} for information on supported values.}

\item{labels}{A \code{\link[base]{vector}} of labels for the treatments.
Should be of \code{\link[base]{length}} \code{D}, containing unique elements.
Defaults to \code{0:(D - 1)}.}

\item{as_matrix}{A \code{\link[base]{logical}} variable indicating whether
the design should be returned as a \code{\link[base]{matrix}}, or a
\code{\link[tibble]{tibble}}. Defaults to \code{T}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
Either a \code{\link[base]{matrix}} if \code{as_matrix = T} (with
rows corresponding to sequences and columns to periods), or a
\code{\link[tibble]{tibble}} if \code{as_matrix = F} (with rows corresponding
to a particular period on a particular sequence). In either case, the
returned object will have class \code{xover_seq}.
}
\description{
Specifies cross-over designs from Pigeon (1984).
}
\details{
\code{seq_pigeon()} supports the specification of designs from
Pigeon (1984). Designs for three through seven treatments (see
\code{D}) are supported, for any chosen treatment labels (see \code{labels}).
In addition, the designs can be returned in \code{\link[base]{matrix}} or
\code{\link[tibble]{tibble}} form (see \code{as_matrix}).

Precisely, for \code{D} equal to three through seven, there are one, four,
six, seven, and two designs available respectively (accessible by setting
\code{selection} equal to one through seven as appropriate). Ultimately, the
\ifelse{html}{\out{(<i>k</i>,<i>j</i>)}}{\eqn{(k,j)}}th element of the
cross-over design matrix corresponds to the treatment a subject on the
ifelse{html}{\out{<i>k</i>}}{\eqn{k}}th sequence would receive in the
\ifelse{html}{\out{<i>j</i>}}{\eqn{j}}th period.
}
\examples{
# Pigeon (1984) designs for four treatments
pigeon1        <- seq_pigeon(D = 4)
pigeon2        <- seq_pigeon(D = 4, selection = 2)
pigeon3        <- seq_pigeon(D = 4, selection = 3)
pigeon4        <- seq_pigeon(D = 4, selection = 4)
# Using different labels
pigeon1_ABCD   <- seq_pigeon(D = 4, labels = LETTERS[1:4])
pigeon2_ABCD   <- seq_pigeon(D = 4, selection = 2, labels = LETTERS[1:4])
pigeon3_ABCD   <- seq_pigeon(D = 4, selection = 3, labels = LETTERS[1:4])
pigeon4_ABCD   <- seq_pigeon(D = 4, selection = 4, labels = LETTERS[1:4])
# Returning in tibble form
pigeon1_tibble <- seq_pigeon(D = 4, as_matrix = F)
pigeon2_tibble <- seq_pigeon(D = 4, selection = 2, as_matrix = F)
pigeon3_tibble <- seq_pigeon(D = 4, selection = 3, as_matrix = F)
pigeon4_tibble <- seq_pigeon(D = 4, selection = 4, as_matrix = F)
}
\references{
Pigeon JG (1985) Residual effects designs for comparing
treatments with a control. \emph{PhD thesis, Temple University}.
}
\author{
Based on data from the \code{\link[Crossover]{Crossover}} package by
Kornelius Rohmeyer.
}
