% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_russell.R
\name{seq_russell}
\alias{seq_russell}
\title{Russell (1991) cross-over design specification}
\usage{
seq_russell(D = 5, labels = 0:(D - 1), as_matrix = T, summary = T)
}
\arguments{
\item{D}{The number of treatments. Must be either five or seven. Defaults to
\code{5}.}

\item{labels}{A \code{\link[base]{vector}} of labels for the treatments.
Should be of \code{\link[base]{length}} \code{D}, containing unique elements.
Defaults to \code{0:(D - 1)}.}

\item{as_matrix}{A \code{\link[base]{logical}} variable indicating whether
the design should be returned as a \code{\link[base]{matrix}}, or a
\code{\link[tibble]{tibble}}. Defaults to \code{T}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
Either a \code{\link[base]{matrix}} if \code{as_matrix = T} (with
rows corresponding to sequences and columns to periods), or a
\code{\link[tibble]{tibble}} if \code{as_matrix = F} (with rows corresponding
to a particular period on a particular sequence). In either case, the
returned object will have class \code{xover_seq}.
}
\description{
Specifies cross-over designs from Russell (1991).
}
\details{
\code{seq_russell()} supports the specification of designs from Russell
(1991). Designs for five or seven treatments (see \code{D}) are supported,
for any chosen treatment labels (see \code{labels}). In addition, the designs
can be returned in \code{\link[base]{matrix}} or \code{\link[tibble]{tibble}}
form (see \code{as_matrix}).

Precisely, the \ifelse{html}{\out{(<i>k</i>,<i>j</i>)}}{\eqn{(k,j)}}th
element of the cross-over design matrix corresponds to the treatment a
subject on the \ifelse{html}{\out{<i>k</i>}}{\eqn{k}}th sequence would
receive in the \ifelse{html}{\out{<i>j</i>}}{\eqn{j}}th period.
}
\examples{
# Russell (1991) design for five treatments
russell        <- seq_russell()
# Using different labels
russell_ABCDE  <- seq_russell(labels = LETTERS[1:5])
# Returning in tibble form
russell_tibble <- seq_russell(as_matrix = F)
}
\references{
Russell KR (1991) The construction of good change-over designs
when there are fewer units than treatments. \emph{Biometrika}
\strong{78:}305-313
}
\author{
Based on data from the \code{\link[Crossover]{Crossover}} package by
Kornelius Rohmeyer.
}
