% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_williams.R
\name{seq_williams}
\alias{seq_williams}
\title{Williams cross-over design specification}
\usage{
seq_williams(D = 2, labels = 0:(D - 1), as_matrix = T, summary = T)
}
\arguments{
\item{D}{The number of treatments. Must be a single
\code{\link[base]{numeric}} integer greater than or equal to two. Defaults to
\code{2}.}

\item{labels}{A \code{\link[base]{vector}} of labels for the treatments.
Should be of \code{\link[base]{length}} \code{D}, containing unique elements.
Defaults to \code{0:(D - 1)}.}

\item{as_matrix}{A \code{\link[base]{logical}} variable indicating whether
the design should be returned as a \code{\link[base]{matrix}}, or a
\code{\link[tibble]{tibble}}. Defaults to \code{T}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
Either a \code{\link[base]{matrix}} if \code{as_matrix = T} (with
rows corresponding to sequences and columns to periods), or a
\code{\link[tibble]{tibble}} if \code{as_matrix = F} (with rows corresponding
to a particular period on a particular sequence). In either case, the
returned object will have class \code{xover_seq}.
}
\description{
Specifies Williams cross-over designs.
}
\details{
\code{seq_williams()} supports the specification of Williams designs.
Sequences for any number of treatments (see \code{D}) are supported, for any
chosen treatment labels (see \code{labels}). In addition, the designs can be
returned in \code{\link[base]{matrix}} or \code{\link[tibble]{tibble}} form
(see \code{as_matrix}).

Precisely, Williams designs are (generalized) Latin squares that are balanced
for first order carryover effects. Generally, carryover balance is achieved
with very few sequences. Ultimately, the
\ifelse{html}{\out{(<i>k</i>,<i>j</i>)}}{\eqn{(k,j)}}th element of the
cross-over design matrix corresponds to the treatment a subject on the
\ifelse{html}{\out{<i>k</i>}}{\eqn{k}}th sequence would receive in the
\ifelse{html}{\out{<i>j</i>}}{\eqn{j}}th period.
}
\examples{
# A Williams design for three treatments
williams        <- seq_williams(D = 3)
# Using different labels
williams_ABC    <- seq_williams(D = 3, labels = LETTERS[1:3])
# Returning in tibble form
williams_tibble <- seq_williams(D = 3, as_matrix = F)
}
\references{
Jones B, Kenward MG (2014) \emph{Design and Analysis of
Cross-Over Trials}. Chapman and Hall: London, 3rd Edition.

Wakeling IN, MacFie HJH (1995) Designing consumer trials balanced
for first and higher orders of carry-over effect when only a subset of k
samples from t may be tested. \emph{Food Qual Prefer} \strong{6:}299-308.

Williams EJ (1949) Experimental designs balanced for the
estimation of residual effects of treatments. \emph{Aust J Sci Res Ser A}
\strong{2:}149-168.
}
\author{
Based on code from the \code{\link[crossdes]{crossdes}} package by
Oliver Sailer.
}
