% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_williams_bib.R
\name{seq_williams_bib}
\alias{seq_williams_bib}
\title{Carryover balanced cross-over design specification}
\usage{
seq_williams_bib(sequences, as_matrix = T, summary = T)
}
\arguments{
\item{sequences}{A balanced incomplete block design, of class
\code{xover_seq}.}

\item{as_matrix}{A \code{\link[base]{logical}} variable indicating whether
the design should be returned as a \code{\link[base]{matrix}}, or a
\code{\link[tibble]{tibble}}. Defaults to \code{T}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
Either a \code{\link[base]{matrix}} if \code{as_matrix = T} (with
rows corresponding to sequences and columns to periods), or a
\code{\link[tibble]{tibble}} if \code{as_matrix = F} (with rows corresponding
to a particular period on a particular sequence). In either case, the
returned object will have class \code{xover_seq}.
}
\description{
Specifies carryover balanced cross-over designs.
}
\details{
\code{seq_williams_bib()} supports the specification of carryover balanced
generalized Youden designs, based on any input balanced incomplete block
(BIB) design (see \code{sequences}). In addition, the designs can be returned
in \code{\link[base]{matrix}} or \code{\link[tibble]{tibble}} form (see
\code{as_matrix}).

Precisely, Patterson (1951) combined BIB designs with Williams designs to get
carryover balanced generalized Youden designs. For each sequence in the input
BIB design, a Williams design is constructed using the treatments in that
sequence. The sequences of the resulting Williams designs are then combined.
Ultimately, the \ifelse{html}{\out{(<i>k</i>,<i>j</i>)}}{\eqn{(k,j)}}th
element of the cross-over design matrix corresponds to the treatment a
subject on the \ifelse{html}{\out{<i>k</i>}}{\eqn{k}}th sequence would
receive in the \ifelse{html}{\out{<i>j</i>}}{\eqn{j}}th period.

Note that the resulting design will only be properly balanced if the input
design is actually a BIB design. A warning message will be printed if this is
not the case.

BIB designs can be generated using \code{\link[xover]{seq_bib}}.
}
\examples{
# Find a BIB design
bib                 <- seq_bib()
# Use this to construct a carryover balanced design
williams_bib        <- seq_williams_bib(bib)
# Returning in tibble form
williams_bib_tibble <- seq_williams_bib(bib, as_matrix = F)
}
\references{
Patterson HD (1951) Change-over trials. \emph{J R Stat Soc B}
\strong{13:}256-271.

Wakeling IN, MacFie HJH (1995) Designing consumer trials balanced
for first and higher orders of carry-over effect when only a subset of k
samples from t may be tested. \emph{Food Qual Prefer} \strong{6:}299-308.

Williams EJ (1949) Experimental designs balanced for the
estimation of residual effects of treatments. \emph{Aust J Sci Res Ser A}
\strong{2:}149-168.
}
\seealso{
\code{\link[crossdes]{seq_bib}},
\code{\link[crossdes]{seq_williams}}.
}
\author{
Based on code from the \code{\link[crossdes]{crossdes}} package by
Oliver Sailer.
}
