% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bin.R
\name{sim_bin}
\alias{sim_bin}
\title{Simulate binary outcome data from a cross-over trial}
\usage{
sim_bin(sequences, n = 1, replicates = 1000, intercept, period,
  treatment, carryover, subject, sigma_b = 1, random = T,
  seed = Sys.time(), summary = T)
}
\arguments{
\item{sequences}{An object of class \code{xover_seq}, describing the
cross-over design for which data will be simulated.}

\item{n}{Either a single \code{\link[base]{numeric}} integer describing the
number of subjects assigned to all of the sequences, or a
\code{\link[base]{numeric}} vector of integers having length equal to the
number of sequences implied by \code{sequences}, with the elements describing
the number of subject assigned to each of the sequences.}

\item{replicates}{A single \code{\link[base]{numeric}} integer describing the
number of replicate datasets to generate.}

\item{intercept}{A single \code{\link[base]{numeric}} describing the
intercept of the GLMM.}

\item{period}{A \code{\link[base]{numeric}} vector of length equal to one
less than the number of periods implied by \code{sequences}, describing the
values of each of the period effects.}

\item{treatment}{A \code{\link[base]{numeric}} vector of length equal to one
less than the number of treatments implied by \code{sequences}, describing
the values of each of the treatment effects.}

\item{carryover}{A \code{\link[base]{numeric}} vector of length equal to one
less than the number of treatments implied by \code{sequences}, describing
the values of each of the carryover effects.}

\item{subject}{A \code{\link[base]{numeric}} vector of length equal to one
less than the number of subjects implied by \code{n} and \code{sequences},
describing the values of each of the subject effects.}

\item{sigma_b}{A single strictly positive \code{\link[base]{numeric}}
describing the value of the between subject standard deviation.}

\item{random}{A \code{\link[base]{logical}} variable indicating whether to
treat the subject effects as random or fixed.}

\item{seed}{A random number seed, to be passed to
\code{\link[base]{set.seed}}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
A \code{\link[base]{list}} or class \code{xover_sim} containing in
particular a \code{\link[base]{list}} \code{sim_data} which holds the
simulated data.
}
\description{
Facilitates the simulation of binary outcome data from a cross-over trial,
for a given cross-over design.
}
\details{
\code{sim_bin()} supports the simulation of binary outcome data from a
cross-over trial. Precisely, a cross-over design of class \code{xover_seq} is
provided (see \code{sequences}). Then, the values of \code{treatment},
\code{period}, and several other input arguments determines the distribution
of the simulated data. See the package vignette for further dettails.
}
\references{
Jones B, Kenward MG (2014) \emph{Design and Analysis of
Cross-Over Trials}. Chapman and Hall: London, 3rd Edition.
}
\seealso{
\code{\link[xover]{as_xover_seq}} for converting designs to class
\code{xover_seq}.
}
