% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_seq_clatworthy.R
\name{summary_seq_clatworthy}
\alias{summary_seq_clatworthy}
\title{Available cross-over designs from Clatworthy (1973)}
\usage{
summary_seq_clatworthy(D, J, K, summary = T)
}
\arguments{
\item{D}{The number of treatments. If specified, must be a single
\code{\link[base]{numeric}} integer greater than or equal to two. Defaults to
missing, which means no subsetting of the available designs is performed
based on the number of treatments.}

\item{J}{The number of periods. If specified, must be a
\code{\link[base]{numeric}} vector, containing integer values greater than or
equal to two. Defaults to missing, which means no subsetting of the available
designs is performed
based on the number of periods.}

\item{K}{The number of sequences. If specified, must be a
\code{\link[base]{numeric}} vector, containing integer values greater than or
equal to two. Defaults to missing, which means no subsetting of the available
designs is performed based on the number of sequences.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether a
summary of the function's progress should be printed to the console. Defaults
to \code{T}.}
}
\value{
A \code{\link[tibble]{tibble}} containing the available designs from
\code{\link[xover]{seq_clatworthy_williams}} for the chosen values of
\code{D}, \code{J}, and \code{K}.
}
\description{
Summarises the cross-over designs available from Clatworthy (1973) via the
function \code{\link[xover]{seq_clatworthy}}.
}
\details{
\code{summary_seq_clatworthy()} supports the determination of which designs
are available from \code{\link[xover]{seq_clatworthy}}. Precisely, values
for the number of treatments (see \code{D}), periods (see \code{J}), and
sequences (see \code{K}) desired in the design is specified. Then, the
function returns which compatible designs are supported by
\code{\link[xover]{seq_clatworthy}}.
}
\examples{
# Available designs for six treatments from Clatworthy (1973)
clatworthy_D6     <- summary_seq_clatworthy(D = 6)
# Available designs for six treatments and three periods
clatworthy_D6J3   <- summary_seq_clatworthy(D = 6, J = 3)
# Available designs for six treatments, three periods, and eight sequences
clatworthy_D6J3K8 <- summary_seq_clatworthy(D = 6, J = 3, K = 8)
# Returning in tibble form
}
\references{
Clatworthy WH, Cameron JM, Speckman JA (1973) Tables of
two-associate-class partially balanced designs. \emph{US Government Printing
Office}.
}
\seealso{
\code{\link[xover]{summary_seq_clatworthy}}.
}
