% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xover.R
\docType{package}
\name{xover}
\alias{xover}
\alias{xover-package}
\title{\strong{xover}: A package for designing and analysing cross-over clinical
trials}
\description{
\strong{xover} provides five main categories of important function:
\code{an}, \code{des}, \code{seq}, \code{sim}, and \code{plot}. Their
respective purpose is as follows:
}
\details{
\itemize{
\item \code{an}: \strong{\emph{An}}alysis of cross-over trial data;
\item \code{des}: Computation of the sample size required by a cross-over
trials (i.e., cross-over trial \strong{\emph{des}}ign);
\item \code{seq}: Specification of cross-over designs (i.e., sets of
\strong{\emph{seq}}uences);
\item \code{sim}: \strong{\emph{Sim}}ulation of data from a cross-over trial;
\item \code{plot}: \strong{\emph{Plot}}ting of cross-over trial data.
}

For an overview of the package's functionality, a good place to start is
\href{https://github.com/mjg211/xover/}{here}. For a more detailed
description, please see the available vignette.

more detail description of the package's functionality, please see the
available vignette.
}
